Changed CSS and JS dirs, minor config cleanups, dropped default theme, setup base for Imprimatur theme, directly integrated littlefoot
This involved moving lots of parts around into something I see as more
logical. There’s no real styling yet, but the nuts and bolts work.

TODO: Social links, Blogroll, Styling

Littlefoot was directly integrated because the gem likes to do wonky
stuff with the CSS and JS files, and I was tired of dealing with that.